add testing framework
tests are based on assembly snippets with testing annotations.

the assembly snippets are assembled into executable binaries. run.sh
executes the binary (currently just in the simulator), then compares
the final processor state to the expected state declared in the
annotation in the assembly file.

the makefile generates a list of all tests by finding all .s files in
tests/, and can assemble and run the tests.