top module for icebreaker board
add a top module to put the system on the icebreaker board.
switch the makefile to synthesize this instead of wb_system.
add a basic cocotb test for this icebreaker_top.

the ROM contains the assembled icebreaker_top.s, which writes
out incrementing values to the rom port (for display on the leds)
about once a second (with nested loops to generate the delay cycles).