add define to control the size of the rom memory
the rom / ram should be updated so that they can use ICESTORM_RAM,
but for now add this to reduce resource usage for testing.

once the bytes per rom is set to the correct value, reducing rom size
with this would have some weird effects if there are more than one rom,
because the valid rom addresses will be sparse.