registered output for rom memory
ice40 block rams have registered outputs, so in
order to use them memories also need to use registered outputs.

switch rom to use a registered output (memory_o).

in order to have the data available in subcycle 3, the high bits
of the 12 bit address need to be used directly from data_i, so
reduce the address register to 8 bits.