switch rom io to input, implement RDR
ram can be used as output, so switch rom io lines to be outputs.
ideally this would be parameterized somehow so it could be different for
each rom instance.

also implement RDR in the cpu_control decoder, and add some test /
testbench hacks to make the tests work for rtl simulation